added LungVolume and LungCapacity psuedocode.
How the Ventilator can work based on the person's lung capacity and its current lung volume. We don't want this machine to overfill the lungs or else the lungs will explode due to too much air. 0.005 is just a placeholder unless we find a correlation between the person's lung capacity/volume and how much air the ventilator transmits.